+++ contracts/BoxV2.sol
@@ -26,7 +26,7 @@ contract BoxV2 is Initializable, BoxManagerRole {
         return _value;
     }
 
-    function version() public pure returns (uint256) {
+    function version() public pure returns (uint16) {
         return 2;
     }
